+++ server/websocket_requests.go
@@ -2,8 +2,12 @@ package server
 
 import (
 	"encoding/json"
+	"errors"
 	"fmt"
+	"strings"
 	"time"
+
+	"github.com/evan-buss/openbooks/irc"
 
 	"github.com/evan-buss/openbooks/core"
 	"github.com/evan-buss/openbooks/util"
@@ -19,6 +23,8 @@ func (server *server) routeMessage(message Request, c *Client) {
 	var obj interface{}
 
 	switch message.MessageType {
+	case CONNECT:
+		obj = new(ConnectionRequest)
 	case SEARCH:
 		obj = new(SearchRequest)
 	case DOWNLOAD:
@@ -37,7 +43,7 @@ func (server *server) routeMessage(message Request, c *Client) {
 
 	switch message.MessageType {
 	case CONNECT:
-		c.startIrcConnection(server)
+		c.startIrcConnection(obj.(*ConnectionRequest), server)
 	case SEARCH:
 		c.sendSearchRequest(obj.(*SearchRequest), server)
 	case DOWNLOAD:
@@ -48,15 +54,23 @@ func (server *server) routeMessage(message Request, c *Client) {
 }
 
 // handle ConnectionRequests and either connect to the server or do nothing
-func (c *Client) startIrcConnection(server *server) {
+func (c *Client) startIrcConnection(request *ConnectionRequest, server *server) {
 	// The IRC connection could be re-used if it is already connected
 	if !c.irc.IsConnected() {
-		err := core.Join(c.irc, server.config.Server, server.config.EnableTLS)
+		err := core.Join(c.irc, request.Address, request.Channel, request.EnableTLS)
 		if err != nil {
 			c.log.Println(err)
-			c.send <- newErrorResponse("Unable to connect to IRC server.")
+			if errors.Is(err, irc.ErrTLSHandshake) {
+				c.send <- newErrorResponse(err.Error())
+			} else {
+				c.send <- newErrorResponse("Unable to connect to IRC server.")
+			}
 			return
 		}
+
+		address := strings.Split(request.Address, ":")[0]
+
+		c.log.Printf("Connected to %s #%s as %s.\n", address, request.Channel, c.irc.Username)
 
 		handler := server.NewIrcEventHandler(c)
 
@@ -74,7 +88,7 @@ func (c *Client) startIrcConnection(server *server) {
 			StatusResponse: StatusResponse{
 				MessageType:      CONNECT,
 				NotificationType: SUCCESS,
-				Title:            "Welcome, connection established.",
+				Title:            fmt.Sprintf("Connection established to %s #%s", address, request.Channel),
 				Detail:           fmt.Sprintf("IRC username %s", c.irc.Username),
 			},
 			Name: c.irc.Username,